requirements.txt
fastapi==0.127.0
uvicorn==0.39.0
starlette==0.49.3
python-multipart==0.0.20
pillow==11.3.0
psycopg2-binary==2.9.11
SQLAlchemy==2.0.45
pydantic==2.12.5
pydantic-settings==2.11.0
python-dotenv==1.2.1
numpy==2.0.2
pandas==2.3.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
requests==2.32.5
